Fix progress bar "flatness" & allow it to visibly complete
Extend the animation duration from 0.9 to 0.8 times the javascript to replace it
(so it finishes before it's replaced & visibly reaches 100% most of the time,
depending on browser performance)

and

Fix the right side being weirdly flat because of a smaller-radius rounding
on the right side than the left. Equalises this to the left which is prettier.